Issue #112: Refine grouping of profiles.
Created context class.
Moved environment and context related functions to context.
Added get_group_by_value() to context.
Added groupby column to database.
Slowest - grouped page now groups using the groupby value.
